+++ Gems/ROS2Sensors/Code/Include/ROS2Sensors/Lidar/LidarTemplate.h
