+++ Gems/ROS2Sensors/Code/Source/Camera/ROS2CameraSensorEditorComponent.h
@@ -15,7 +15,7 @@
 #include <ROS2/Frame/NamespaceConfiguration.h>
 #include <ROS2/Frame/ROS2Transform.h>
 #include <ROS2Sensors/Camera/CameraCalibrationRequestBus.h>
-#include <ROS2Sensors/Configuration/CameraSensorConfiguration.h>
+#include <ROS2Sensors/Camera/CameraSensorConfiguration.h>
 #include <ROS2Sensors/Sensor/SensorConfiguration.h>
 
 namespace ROS2
