+++ Gems/ROS2Sensors/Code/Source/Imu/ImuSensorConfiguration.cpp
@@ -8,7 +8,7 @@
 
 #include <AzCore/Serialization/EditContext.h>
 #include <AzCore/Serialization/EditContextConstants.inl>
-#include <ROS2Sensors/Configuration/ImuSensorConfiguration.h>
+#include <ROS2Sensors/Imu/ImuSensorConfiguration.h>
 
 namespace ROS2
 {
