+++ Gems/ROS2Sensors/Code/Source/Odometry/ROS2WheelOdometry.cpp
@@ -10,7 +10,7 @@
 #include <ROS2/Utilities/ROS2Conversions.h>
 #include <ROS2/Utilities/ROS2Names.h>
 #include <ROS2/VehicleDynamics/VehicleInputControlBus.h>
-#include <ROS2Sensors/Configuration/ROS2OdometryCovariance.h>
+#include <ROS2Sensors/Odometry/ROS2OdometryCovariance.h>
 
 namespace ROS2
 {
