+++ Gems/ROS2Sensors/Code/Source/Odometry/ROS2WheelOdometry.h
@@ -9,7 +9,7 @@
 
 #include <AzCore/Math/Transform.h>
 #include <ROS2/Sensor/Events/PhysicsBasedSource.h>
-#include <ROS2Sensors/Configuration/ROS2WheelOdometryConfiguration.h>
+#include <ROS2Sensors/Odometry/ROS2WheelOdometryConfiguration.h>
 #include <ROS2Sensors/Sensor/ROS2SensorComponentBase.h>
 #include <nav_msgs/msg/odometry.hpp>
 #include <rclcpp/publisher.hpp>
