+++ Gems/ROS2Sensors/Code/Source/ROS2SensorsModuleInterface.cpp
@@ -29,8 +29,6 @@
 #include <Odometry/ROS2OdometrySensorComponent.h>
 #include <Odometry/ROS2WheelOdometry.h>
 
-#include <TestComponent.h>
-
 namespace ROS2Sensors
 {
     AZ_TYPE_INFO_WITH_NAME_IMPL(ROS2SensorsModuleInterface, "ROS2SensorsModuleInterface", ROS2SensorsModuleInterfaceTypeId);
@@ -47,8 +45,6 @@ namespace ROS2Sensors
             m_descriptors.end(),
             {
                 ROS2SensorsSystemComponent::CreateDescriptor(),
-                // ROS2::ROS2SensorComponentBase<ROS2::TickBasedSource>::CreateDescriptor(),
-                // ROS2::ROS2SensorComponentBase<ROS2::PhysicsBasedSource>::CreateDescriptor(),
                 ROS2::ROS2CameraSensorComponent::CreateDescriptor(),
                 ROS2::ROS2SystemCameraComponent::CreateDescriptor(),
                 ROS2::ROS2ImageEncodingConversionComponent::CreateDescriptor(),
@@ -61,8 +57,6 @@ namespace ROS2Sensors
                 ROS2::LidarRegistrarSystemComponent::CreateDescriptor(),
                 ROS2::ROS2OdometrySensorComponent::CreateDescriptor(),
                 ROS2::ROS2WheelOdometryComponent::CreateDescriptor(),
-
-                TestNamespace::TestComponent::CreateDescriptor(),
             });
     }
